LPS-71164 Move class to util folder
